// web/src/Enums/document.enum.ts
export enum DocType {
  DESIGN_DOCUMENT = '0',
  METHODOLOGY_DOCUMENT = '1',
  MONITORING_REPORT = '2',
  VERIFICATION_REPORT = '3',
  NO_OBJECTION_LETTER = '4',
  AUTHORISATION_LETTER = '5',
  ANNUAL_REPORT = '6',
  ENVIRONMENTAL_IMPACT_ASSESSMENT = '7',
}

export enum DocumentStatus {
  PENDING = 'Pending',
  ACCEPTED = 'Accepted',
  REJECTED = 'Rejected',
}